+++ CHANGELOG.md
@@ -10,6 +10,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 ### Removed
 
+## [1.0.0-beta.4] - 2018-12-22
+### Added
+- Add "notifications" feature. Add "notify" property to configuration in order to map an ability state into a HomeKit notification.
+
+### Changed
+- Change accessories default name. Add service name after accesory name.
+
 ## [1.0.0-beta.3] - 2018-12-18
 ### Added
 - Use controller services and configs instead of abilities to map HomeKit accessories